+++ Control/Joint/Effects/Either.hs
@@ -4,7 +4,7 @@ import Control.Joint.Operators ((<$$>), (<**>))
 import Control.Joint.Abilities.Interpreted (Interpreted (Primary, run))
 import Control.Joint.Abilities.Transformer (Transformer (build, unite), Schema, (:>) (T))
 import Control.Joint.Abilities.Adaptable (Adaptable (adapt))
-import Control.Joint.Schemes (UT (UT))
+import Control.Joint.Schemes (UT (UT), type (<.:>))
 
 instance Interpreted (Either e) where
 	type Primary (Either e) a = Either e a
@@ -16,14 +16,14 @@ instance Transformer (Either e) where
 	build x = T . UT . pure $ x
 	unite = T . UT
 
-instance Functor u => Functor (UT (Either e) u) where
+instance Functor u => Functor (Either e <.:> u) where
 	fmap f (UT x) = UT $ f <$$> x
 
-instance Applicative u => Applicative (UT (Either e) u) where
+instance Applicative u => Applicative (Either e <.:> u) where
 	pure = UT . pure . pure
 	UT f <*> UT x = UT $ f <**> x
 
-instance (Applicative u, Monad u) => Monad (UT (Either e) u) where
+instance (Applicative u, Monad u) => Monad (Either e <.:> u) where
 	UT x >>= f = UT $ x >>= either (pure . Left) (run . f)
 
 type Failable e = Adaptable (Either e)