+++ Control/Joint/Schemes/TU.hs
@@ -9,7 +9,9 @@ import Control.Joint.Abilities (Interpreted (Primary, run))
 
 newtype TU t u a = TU (t :. u := a)
 
-instance Interpreted (TU t u) where
+type (<:.>) = TU
+
+instance Interpreted (t <:.> u) where
 	type Primary (TU t u) a = t :. u := a
 	run (TU x) = x