+++ CHANGELOG.md
@@ -63,3 +63,4 @@
 * Define `Concepts` umbrella module with `Lens` submodule
 * Define `zoom` method in `Concepts.Lens` module
 * Define `_1` and `_2` lenses for tuples
+* Define `<:.>`, `<.:>` and `<:<.>:>` type operators for `TU`, `UT` and `TUT`